% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy_help}
\alias{accuracy_help}
\title{Calculate accuracy given a single quantile pair}
\usage{
accuracy_help(fcst, obs, pair)
}
\arguments{
\item{fcst}{A forecast object.}

\item{obs}{An observations data frame.}

\item{pair}{A valid quantile pair}
}
\value{
A data frame with \code{time}, \code{val_obs}, and \code{score} columns
}
\description{
Helper for \code{accuracy()}.
Calculates accuracy for a single quantile pair and returns the unsummarized result.
}
\examples{
# See `?accuracy`
}
