% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{all_of}
\alias{all_of}
\title{Check everything in a vector/list}
\usage{
all_of(v, f)
}
\arguments{
\item{v}{A list or vector, presumably a column in a data frame.}

\item{f}{A predicate function (accepts one argument and returns a boolean).}
}
\value{
TRUE if everything in \code{v} passes \code{f()}, FALSE otherwise
}
\description{
\code{all_of()} checks that all elements of a given vector or list pass a given predicate function.
Useful for checking types.
}
\examples{
# returns TRUE
casteval:::all_of(list(1, 2.5, 1.4e7, -10, 5L, Inf, NaN), is.numeric)

# returns FALSE
casteval:::all_of(list(1, 2, 3), function(x) {x < 3})

# NA is by default a logical value, not a numeric one. Returns FALSE
casteval:::all_of(list(1, NA), is.numeric)

# NA gets coerced to NA_real_, because it's in a numeric vector. Returns TRUE
casteval:::all_of(c(1,NA), is.numeric)
}
