% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{column_all}
\alias{column_all}
\title{Check everything in a column}
\usage{
column_all(col, f)
}
\arguments{
\item{col}{A list or vector, presumably a column in a data frame.}

\item{f}{A predicate function (accepts one argument and returns a boolean).}
}
\value{
TRUE if everything in col passes \code{f()}, FALSE otherwise
}
\description{
\code{column_all()} checks that all elements of a given column pass a given predicate function.
Useful for checking types.
}
\examples{
# returns TRUE
casteval:::column_all(list(1, 2.5, 1.4e7, -10, 5L, Inf, NaN), is.numeric)

# returns FALSE
casteval:::column_all(list(1, 2, 3), function(x) {x < 3})

# NA is by default a logical value, not a numeric one. Returns FALSE
casteval:::column_all(list(1, NA), is.numeric)

# NA gets coerced to NA_real_, because it's in a numeric vector. Returns TRUE
column_all(c(1,NA), is.numeric)
}
