% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ensemble.R
\name{combine_ensemble}
\alias{combine_ensemble}
\title{Combine a group of forecasts into one}
\usage{
combine_ensemble(...)
}
\arguments{
\item{...}{The forecasts (individual realizations) to be combined.
They should have the same time type and all contain raw values only.
Forecasts should be formatted as named lists, such as the ones \code{create_forecast()} returns.}
}
\value{
A named list representing the resulting forecast.
}
\description{
User-facing wrapper for \code{combine_data_frames()}.
Converts multiple forecasts into a single forecast.
}
\examples{
#TBD
}
