% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_frames.R
\name{combine_two_data_frames}
\alias{combine_two_data_frames}
\title{Combine two forecast data frames into one}
\usage{
combine_two_data_frames(df1, df2)
}
\arguments{
\item{df1}{The first data frame.}

\item{df2}{The second data frame.}
}
\value{
The combined data frame.
}
\description{
Helper for \code{combine_data_frames()}.
See \code{?combine_data_frames()} for details.
}
\examples{
# full_join() is used to make times line up
casteval:::combine_two_data_frames(
  data.frame(time=c(1,2,3), raw=c(10, 11, 12)),
  data.frame(time=c(3,2,1), raw=c(20, 21, 22))
)
}
