% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast_ensemble}
\alias{create_forecast_ensemble}
\title{Create forecast from time vector and ensemble of realizations}
\usage{
create_forecast_ensemble(time, vals)
}
\arguments{
\item{time}{A vector of times}

\item{vals}{A list of vectors of values.
Each vector corresponds to a realization which will be assigned a simulation number.}
}
\value{
A forecast data frame
}
\description{
Helper for \code{create_forecast()}.
}
\examples{
# See `?create_forecast`
}
