% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast_multiple}
\alias{create_forecast_multiple}
\title{Create forecast from multiple data frames}
\usage{
create_forecast_multiple(dfs, name, forecast_time)
}
\arguments{
\item{dfs}{A list of data frames}

\item{name}{A string}

\item{forecast_time}{A number, date, or date-time}
}
\value{
A forecast object
}
\description{
Helper for \code{create_forecast()}.
}
\examples{
# See `create_forecast()`
}
