% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast_single}
\alias{create_forecast_single}
\title{Create forecast from single data frame}
\usage{
create_forecast_single(df, name, forecast_time)
}
\arguments{
\item{df}{A data frame}

\item{name}{A string}

\item{forecast_time}{A number, date, or date-time}
}
\value{
A forecast object
}
\description{
Helper for \code{create_forecast()}.
}
\examples{
# See `create_forecast()`
}
