% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020df}
\alias{denmark2020df}
\title{Denmark 2020 COVID-19 forecast data frame}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75000 rows and 3 columns.
}
\source{
\url{https://www.nature.com/articles/s41567-020-01121-y}

\url{https://github.com/jonassjuul/curvestat}
}
\usage{
denmark2020df
}
\description{
The data frame that belongs to \code{denmark2020fc}, exported for convenience.
}
\seealso{
\link{denmark2020fc}, \link{denmark2020obs}, \link{denmark2020ens}
}
\keyword{datasets}
