% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020ens}
\alias{denmark2020ens}
\title{Denmark 2020 COVID-19 forecast ensemble}
\format{
An object of class \code{list} of length 500.
}
\source{
\url{https://www.nature.com/articles/s41567-020-01121-y}

\url{https://github.com/jonassjuul/curvestat}
}
\usage{
denmark2020ens
}
\description{
The data used to create \code{denmark2020fc}, as a list of realization vectors.
}
\seealso{
\link{denmark2020fc}, \link{denmark2020df}, \link{denmark2020obs}
}
\keyword{datasets}
