% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020obs}
\alias{denmark2020obs}
\title{Denmark 2020 COVID-19 observations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 204 rows and 2 columns.
}
\usage{
denmark2020obs
}
\description{
An observations data frame containing made-up daily hospitalization data for demonstration purposes.
Meant to be used in conjunction with \code{denmark2020fc} and \code{denmark2020ens}.
}
\seealso{
\link{denmark2020fc}, \link{denmark2020df}, \link{denmark2020ens}
}
\keyword{datasets}
