% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{filter_forecast_time}
\alias{filter_forecast_time}
\title{Isolate projected values from fit values}
\usage{
filter_forecast_time(df, forecast_time)
}
\arguments{
\item{df}{The forecast data frame. See \code{?create_forecast} for details}

\item{forecast_time}{Either NULL or a time of the same type as the values in \code{dat$time}.
If non-NULL, all rows with time prior to \code{forecast_time} will be removed from \code{dat}.}
}
\value{
The filtered data frame (no change if \code{forecast_time} NULL)
}
\description{
Removes rows from forecast data frame which should not be scored,
i.e. the data generated for dates/times prior to when the forecast was created.
}
\examples{
# data.frame(time=5:10, val=15:20)
casteval:::filter_forecast_time(data.frame(time=1:10, val=11:20), 5)

# unchanged
casteval:::filter_forecast_time(data.frame(time=1:10, val=11:20), NULL)
}
