% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_format}
\alias{get_format}
\title{Get data frame format}
\usage{
get_format(df)
}
\arguments{
\item{df}{A data frame. It should have a valid time column and a valid combination of data columns.}
}
\value{
A named list containing a \code{time_type} (a string containing the type of the values in the time column)
and \code{data_types} (a character vector containing the types of data columns present)
}
\description{
Helper function for \code{create_forecast()}.
\code{get_format()} inspects the columns of a data frame to determine how it's formatted.
Also performs input validation on the contents of the data frame.
In particular, it checks that:
\itemize{
\item The data frame isn't empty
\item The \code{time} column is present
\item The \code{time} column contains only dates, only date-times, or only numeric values
\item If raw values are provided, mean and/or quantile values are not also provided
\item At least one data column is present
\item All data columns contain numeric values only
\item Quantile percentages are valid
}
}
\examples{
# numeric times, raw data
casteval:::get_format(data.frame(time=1:3, raw=4:6))
# dates, mean data
casteval:::get_format(
  data.frame(time=c(lubridate::ymd("2024-01-01"), lubridate::ymd("2024-01-02")), mean=1:2))
# dates-times, quantile data (median)
casteval:::get_format(
  data.frame(
    time=c(lubridate::ymd_hms("2024-01-01_01:01:01"), lubridate::ymd_hms("2024-01-02_02:02:02")),
    quant_50=1:2))
# numeric times, mean-and-quantiles data
casteval:::get_format(data.frame(time=1:3, quant_5=4:6, quant_95=7:9, mean=10:12))
}
