% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{get_group_cols}
\alias{get_group_cols}
\title{Get group column names}
\usage{
get_group_cols(df)
}
\arguments{
\item{df}{A forecast data frame}
}
\value{
A character vector of column names
}
\description{
Get group column names of a given forecast data frame
}
\examples{
casteval:::get_group_cols(data.frame(time=1,val=2,grp_scenario=3,grp_variable=4))
}
