% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{get_group_names}
\alias{get_group_names}
\title{Get group names of forecast data}
\usage{
get_group_names(df)
}
\arguments{
\item{df}{A forecast data frame}
}
\value{
A character vector of names
}
\description{
Gets the names of groups in a forecast by inspecting its columns.
}
\examples{
# character(0)
casteval:::get_group_names(data.frame(time=1))

# c("variable", "scenario")
casteval:::get_group_names(data.frame(time=1, grp_variable=2, grp_scenario=3))
}
