% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_obs_format}
\alias{get_obs_format}
\title{Get observations format}
\usage{
get_obs_format(obs)
}
\arguments{
\item{obs}{Data frame containing observations. In order to pass input validation it should:
\itemize{
\item Be non-empty
\item Contain a \code{time} column with a consistent and valid time type
\item Contain a numeric \code{raw} column with only one value per row
}}
}
\value{
The time type, as a string.
}
\description{
Consumes a data frame, performs input validation, and returns the time type.
Essentially a wrapper around \code{get_format()} with stricter requirements.
}
\examples{
# "numeric"
casteval:::get_obs_format(data.frame(time=1:3, raw=4:6))

# "date"
casteval:::get_obs_format(data.frame(time=lubridate::ymd("2024-01-01"), raw=10))

# raw not present
try(casteval:::get_obs_format(data.frame(time=1:3, mean=4:6)))

# multiple raw realizations
try(casteval:::get_obs_format(dplyr::tibble(time=1:3, raw=list(1:2, 3:4, 5:6))))
}
