% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{get_quant_col}
\alias{get_quant_col}
\title{Get quantile column from data frame}
\usage{
get_quant_col(df, num)
}
\arguments{
\item{df}{The data frame. Should contain a "val_q*" column corresponding to \code{num}}

\item{num}{A number between 0 and 100 inclusive.}
}
\value{
The \code{num} quantile column in \code{df}.
}
\description{
Given a data frame and a quantile number, returns the corresponding
column from the data frame.
}
\examples{
# returns 4:6
casteval:::get_quant_col(data.frame(time=1:3, val_q50=4:6), 50)
}
