% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{get_quant_percentages}
\alias{get_quant_percentages}
\title{Get quantile numbers from column names}
\usage{
get_quant_percentages(df)
}
\arguments{
\item{df}{The data frame,
presumably containing columns named "val_q" followed by a number from 0 to 100.}
}
\value{
A numeric vector containing the quantiles, sorted in increasing order.
}
\description{
Given a (forecast) data frame,
inspects the column names and returns the quantile numbers.
This function does no input validation.
}
\examples{
# returns c(2.5, 50, 97.5)
casteval:::get_quant_percentages(
  data.frame(time=1:3, val_q2.5=4:6, val_q50=7:9, val_q97.5=10:12)
)
}
