% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Get quantile numbers from column names}
\usage{
get_quantiles(df)
}
\arguments{
\item{df}{The data frame,
presumably containing columns named "quant_" followed by a number from 0 to 100.}
}
\value{
A numeric vector containing the quantiles, sorted in increasing order.
}
\description{
Given a (forecast) data frame,
inspects the column names and returns the quantile numbers.
This function does no input validation.
}
\examples{
# returns c(2.5, 50, 97.5)
casteval:::get_quantiles(data.frame(time=1:3, quant_2.5=4:6, quant_50=7:9, quant_97.5=10:12))
}
