% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score.R
\name{get_specified_time}
\alias{get_specified_time}
\title{Get relative/absolute time specified by user}
\usage{
get_specified_time(fcst, at = NULL, after = NULL)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{at}{(Optional) See \code{?log_score}}

\item{after}{(Optional) See \code{?log_score}}
}
\value{
A time of the same type as those in \code{fcst}
}
\description{
Get the time according to the \code{at}/\code{after} params passed by the user.
Helper for \code{log_score()} and \code{plot_KDE()}.
}
\examples{
fc <- create_forecast(data.frame(time=1:3, val=4:6), forecast_time=2)
# 1
casteval:::get_specified_time(fc, at=1)

# 2+1=3
casteval:::get_specified_time(fc, after=1)
}
