% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_time_point}
\alias{get_time_point}
\title{Get a row of a date frame for a given time}
\usage{
get_time_point(df, t)
}
\arguments{
\item{df}{A data frame with a \code{time} column.}

\item{t}{A time.}
}
\value{
A data frame with one row, where the time equals \code{t}.
Raises error if there isn't exactly one row which satisfies this.
}
\description{
Given a forecast/observations data frame and a time,
isolate the row with the given time.
}
\examples{
# no rows with time==4
try(casteval:::get_time_point(data.frame(time=1:3,val=4:6), 4))

# 2 rows with time==2
try(casteval:::get_time_point(data.frame(time=c(1,2,2,3), val=4:7), 2))

# data.frame(time=2, raw=5, mean=8)
casteval:::get_time_point(data.frame(time=1:3, val=4:6, val_mean=7:9), 2)
}
