% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_time_type}
\alias{get_time_type}
\title{Get type of time column}
\usage{
get_time_type(timecol)
}
\arguments{
\item{timecol}{A vector or list, presumably a column in a data frame.
Its contents should all be of the same suppotred type (date, date-time, or numeric)}
}
\value{
A string describing the type of the time column
}
\description{
Helper function for \code{get_format()}.
Inspects a time column from a data frame to determine the type of its contents.
Also checks that the contents are all of the same supported type.
}
\examples{
# TBD
}
