% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_confidence_intervals}
\alias{graph_confidence_intervals}
\title{Graph forecast confidence intervals}
\usage{
graph_confidence_intervals(graph = NULL, fcst, conf = NULL)
}
\arguments{
\item{graph}{A ggplot object, or \code{NULL}. If not \code{NULL},
it will be added to using \code{ggplot2}'s \code{+} operator.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{conf}{A numeric vector containing the confidence intervals to be displayed, as percentages.
If \code{NULL}, the confidence intervals will be inferred from the quantiles present in the forecast data frame.}
}
\value{
A ggplot object.
}
\description{
Graph shaded bands indicating confidence intervals of a forecast.
}
\examples{
#TODO
}
