% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_ensemble.R
\name{graph_ensemble}
\alias{graph_ensemble}
\title{Graph ensemble of forecast realizations}
\usage{
graph_ensemble(graph = NULL, fcst)
}
\arguments{
\item{graph}{A ggplot object, or \code{NULL}. If not \code{NULL},
it will be added to using \code{ggplot2}'s \code{+} operator.}

\item{fcst}{The forecast object (see the output of \code{create_forecast()})}
}
\value{
A ggplot object.
}
\description{
Given a forecast with raw realizations, generate a graph displaying all of them.
}
\examples{
#TODO
}
