% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_forecasts}
\alias{graph_forecasts}
\title{Graph one or more forecasts}
\usage{
graph_forecasts(fcsts, obs = NULL, confs = NULL, score = NULL)
}
\arguments{
\item{fcsts}{A single forecast object, or a list of forecast objects (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame.
If provided, they will be overlaid over the forecast(s) as points.}

\item{confs}{(Optional) A vector of numbers from 0 to 100.
The corresponding confidence intervals will be displayed in the resulting graph(s)}

\item{score}{(Optional) A scoring function.
The function will be used to score the \code{obs} against the forecast(s).}
}
\value{
desc
}
\description{
Graph one or more forecasts, along with corresponding observations, quantiles, etc.
}
\examples{
#TODO
}
