% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_observations.R
\name{graph_observations}
\alias{graph_observations}
\title{Intelligently graph observations}
\usage{
graph_observations(graph = NULL, obs)
}
\arguments{
\item{graph}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{obs}{An observations data frame. If it contains a \code{score} column,
the observation points will be color-coded according to the scores.}
}
\value{
A ggplot object.
}
\description{
Graph observation points. If provided with a forecast,
the points can be color-coded to convey information about score, accuracy, etc.
}
\examples{
fc <- create_forecast(dplyr::tibble(
  time=1:3,
  raw=list(4:6, 7:9, 10:12)
))
obs <- data.frame(time=1:3, obs=c(5,9,13))

# graph observations on their own
casteval:::graph_observations(NULL, obs)

# graph observations alongside forecast data
casteval:::graph_observations(casteval:::graph_ensemble(NULL, fc), obs)

# graph observations alongside forecast data, and color-code by score
casteval:::graph_observations(casteval:::graph_ensemble(NULL, fc), neglog(fc, obs))
}
