% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_quantiles}
\alias{graph_quantiles}
\title{Graph forecast quantiles}
\usage{
graph_quantiles(graph = NULL, fcst, quants = NULL)
}
\arguments{
\item{graph}{A ggplot object, or \code{NULL}. If not \code{NULL},
it will be added to using \code{ggplot2}'s \code{+} operator.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{quants}{A numeric vector containing the quantiles to be graphed, as percentages.
If \code{NULL}, all quantiles present in the forecast data frame will be graphed.}
}
\value{
A ggplot object.
}
\description{
Graph lines indicating quantiles of a forecast.
}
\examples{
#TODO
}
