% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupex}
\alias{groupex}
\title{Grouping example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 6 columns.
}
\usage{
groupex
}
\description{
An example data frame for demonstrating \code{{casteval}}'s grouping functionality.
}
\keyword{datasets}
