% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{has_groups}
\alias{has_groups}
\title{Does the forecast data have groups?}
\usage{
has_groups(df)
}
\arguments{
\item{df}{A forecast data frame}
}
\value{
\code{TRUE} if there are groups, \code{FALSE} otherwise
}
\description{
Predicate function which determines whether \code{df} contains \verb{grp_*} columns
}
\examples{
casteval:::has_groups(data.frame(time=1,val=2))

casteval:::has_groups(data.frame(time=1, val=2, grp_scenario=3))
}
