% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Integer breaks on ggplot2 axes}
\source{
https://gist.github.com/jhrcook/eb7b63cc57c683a6eb4986c4107a88ec
}
\usage{
integer_breaks(n = 5, ...)
}
\description{
A modified version of scales::pretty_breaks() for only integer-valued ticks
}
