% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_forecast}
\alias{is_forecast}
\title{Is it a forecast?}
\usage{
is_forecast(fcst)
}
\arguments{
\item{fcst}{An R object}
}
\value{
\code{TRUE} if \code{fcst} is a forecast object, \code{FALSE} otherwise
}
\description{
Determines whether a given R object is a forecast object or not.
}
\details{
An R object is a forecast object if:
\itemize{
\item It is a list (and also not a data frame)
\item It has a \code{name} field (may be \code{NULL})
\item It has a \code{forecast_time} field (may be \code{NULL})
\item It has a \code{data} field
}

\code{is_forecast()} does NO further input validation beyond this.
For a version of the function that does input validation, see \code{is_valid_forecast()}.
}
\examples{
# FALSE
casteval:::is_forecast(5)
casteval:::is_forecast(data.frame(time=1, val=2))

# TRUE
casteval:::is_forecast(list(name=NULL, forecast_time=NULL, data=NULL))
}
