% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_forecast}
\alias{is_valid_forecast}
\title{Is it a valid forecast?}
\usage{
is_valid_forecast(fcst)
}
\arguments{
\item{fcst}{An R object}
}
\value{
\code{TRUE} if \code{fcst} passes \code{validate_forecast()} without any errors,
\code{FALSE} otherwise
}
\description{
A stricter version of \code{is_forecast()} which validates
its input using \code{validate_forecast()}.
}
\examples{
#TODO
}
