% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{join_fcst_obs}
\alias{join_fcst_obs}
\title{Join a forecast and observations into a single data frame}
\usage{
join_fcst_obs(df, obs, na.rm = FALSE)
}
\arguments{
\item{df}{The forecast data frame.}

\item{obs}{The observations data frame.}

\item{na.rm}{A boolean. Defaults to FALSE. If FALSE, then an error will be raised when
observations are missing for any forecast time points.
If TRUE, any time points with missing observations will be removed from the data frame.}
}
\value{
The forecast data frame with an additional \code{obs} column containing observations.
}
\description{
Add an observations column to the forecast data frame containing
observations for each time point.
}
\examples{
# data.frame(time=1:3, raw=4:6, obs=8:10)
casteval:::join_fcst_obs(data.frame(time=1:3, raw=4:6), data.frame(time=0:4, obs=7:11))

# remove rows with missing observations 
# data.frame(time=3, quant_50=6, obs=7)
casteval:::join_fcst_obs(
  data.frame(time=1:3, quant_50=4:6),
  data.frame(time=2:3, obs=c(NA,7)),
  na.rm=TRUE
)

# default behaviour is to error if observations are missing
try(casteval:::join_fcst_obs(
  data.frame(time=1:3, quant_50=4:6),
  data.frame(time=2:3, obs=c(NA,7))
))
}
