% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score_approx_normal.R
\name{log_score_approx_normal}
\alias{log_score_approx_normal}
\title{Logarithmic score for forecast assuming a normal distribution
given paired quantiles.}
\usage{
log_score_approx_normal(fcst, obs, summarize = TRUE, quantpair.choice = 1)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{An observations data frame.}

\item{summarize}{A boolean, defaults to TRUE. If TRUE, a single number will be returned as the score for the forecast.
If FALSE, a data frame with columns named \code{time}, \code{val_obs}, and \code{score} will be returned,
containing the scores for each individual time point.
This can be used by plotting functions to colour-code observations, for example.}

\item{quantpair.choice}{(Optional) Integer. If several quantiles pairs
are present in the forecast object, determine which one to use to
calculate the mean and standard deviation of the implied normal distribution.}
}
\value{
If summarize is \code{FALSE}, a data frame containing times,
observations, scores, quantiles used, and implied mean and standard
deviation of the normal distribution. Otherwise, returns the mean of
all scores at every times.
}
\description{
Given a forecast object having at least one paired quantiles,
assume that the forecasted values are distributed according to
a normal distribution with the same quantile pair.
The logarithmic score is then calculated following this
normality assumption.
}
\examples{
fcst <- create_forecast(
dplyr::tibble(
  time    = 1:3,
  val_q5  = seq(1,8, length.out = 3),
  val_q10 = 9:11,
  val_q25 = 12:14,
  val_q50 = 100:102,
  val_q75 = 200:202,
  val_q90 = 203:205,
  val_q95 = seq(210,220, length.out = 3)
))

obs <- data.frame(time=1:3, val_obs=c(105,101,300))

s = log_score_approx_normal(fcst, obs,
                            summarize = F,
                            quantpair.choice = 1)
print(s)

s = log_score_approx_normal(fcst, obs,
                            summarize = F,
                            quantpair.choice = 2)
print(s)

s = log_score_approx_normal(fcst, obs,
                            summarize = TRUE,
                            quantpair.choice = 3)
print(s)
}
