% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score.R
\name{make_log_score}
\alias{make_log_score}
\title{\code{log_score()} function factory}
\usage{
make_log_score(at = NULL, after = NULL, bw = NULL)
}
\arguments{
\item{at}{See \code{?log_score}}

\item{after}{See \code{?log_score}}

\item{bw}{See \code{?log_score}}
}
\value{
A function wrapping \code{log_score()} with the given parameters
}
\description{
Create a function wrapping \code{log_score()} with specified parameters.
}
\examples{
dat <- c(1.65333590, -0.45354373, -0.72833227, -0.57932896,  0.22393422)
fc <- create_forecast(dplyr::tibble(time=rep(1,5), val=dat), forecast_time=1)
obs <- data.frame(time=1, val_obs=0)

logs <- make_log_score(at=1)
logs(fc, obs)

logs2 <- make_log_score(after=0, bw=0.3)
logs2(fc, obs)
}
