% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neglog.R
\name{neglog}
\alias{neglog}
\title{Get negative-log score for forecast}
\usage{
neglog(fcst, obs)
}
\arguments{
\item{fcst}{The forecast (see \code{create_forecast()} output).}

\item{obs}{The observations data frame.}
}
\value{
A data frame containing times, raw data, observations, and scores for those times.
}
\description{
Given a forecast and set of observations,
compute the negative-log score for every time point.
Uses a Kernel Density Estimation (KDE) to interpolate the density
at the observation point.
}
\examples{
#TODO
}
