% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{pair_quantiles}
\alias{pair_quantiles}
\title{Pair up matching quantiles}
\usage{
pair_quantiles(quants)
}
\arguments{
\item{quants}{A vector of distinct numbers between 0 and 100.}
}
\value{
A named list with two elements: \code{paired}, a list of pairs of quantiles,
and \code{unpaired}, a vector of all the leftover quantiles.
}
\description{
Given a list of quantiles, pair together the ones symmetrical around the median.
For example, the 2.5\% quantile and 97.5\% quantile go together.
The 50\% quantile will never be paired.
}
\examples{
# list(paired=list(c(15,85), c(30,70)), unpaired=c(0,10,20,50,51,60,75))
casteval:::pair_quantiles(c(0, 10, 15, 20, 30, 50, 51, 60, 70, 75, 85))
}
