% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{parse_quant_pairs}
\alias{parse_quant_pairs}
\title{Parse quantile pair(s)}
\usage{
parse_quant_pairs(quant_pairs, df, allow_empty = FALSE)
}
\arguments{
\item{quant_pairs}{A list of pairs, a single pair, or NULL}

\item{df}{A forecast data frame}

\item{allow_empty}{If \code{FALSE}, an error will be raised when no quantile pairs can be inferred.
If \code{TRUE}, an empty list will be returned silently. Defaults to \code{FALSE}.}
}
\value{
A list of pairs, either taken from \code{quant_pairs} or inferred from \code{df}
}
\description{
Helper function for funtions that accept a \code{quant_pairs} argument.
Validates/formats the given pairs, or infers the quantile pairs from forecast data.
}
\examples{
# infer from forecast
casteval:::parse_quant_pairs(
  NULL,
  data.frame(time=1, val_q10=2, val_q25=3, val_50=4, val_q75=5, val_q90=6)
)

# single pair
casteval:::parse_quant_pairs(
  c(2.5,97.5),
  data.frame(time=1, val=1)
)

# multiple pairs
casteval:::parse_quant_pairs(
  list(c(2.5,97.5), c(25,75)),
  data.frame(time=1, val=1)
)

# allow empty
casteval:::parse_quant_pairs(
  NULL,
  data.frame(time=1:3, val_q50=4:6),
  allow_empty=TRUE
)

casteval:::parse_quant_pairs(
  list(),
  data.frame(time=1, val=1),
  allow_empty=TRUE
)
}
