% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble.R
\name{plot_ensemble}
\alias{plot_ensemble}
\title{Plot ensemble of forecast realizations}
\usage{
plot_ensemble(plt = NULL, fcst, alpha = 0.3, colour = "black")
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{alpha}{(Optional) The alpha value to be passed to \code{ggplot2}.}

\item{colour}{(Optional) The colour to be passed to \code{ggplot2}.}
}
\value{
A ggplot object.
}
\description{
Given a forecast with raw realizations, generate a plot displaying all of them.
If the forecast provides simulation numbers (\code{sim} column), the ensemble of curves will be plotted.
Otherwise, the data will be plotted as points.
}
\examples{
plot_ensemble(
  NULL,
  create_forecast(dplyr::tibble(time=rep(1:3,each=3), sim=rep(1:3,3), val=c(4:12))
))

NULL |> plot_ensemble(create_forecast(data.frame(
  time=lubridate::as_datetime(c(0,20000,100000)),
  val=c(20,30,40)
)))
}
