% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot a forecast}
\usage{
plot_forecast(fcst, obs = NULL, quant_pairs = NULL, score = NULL)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame.
If provided, they will be overlaid over the forecast as points.}

\item{quant_pairs}{(Optional) A list of pairs of numbers between 0 and 100,
or a single pair of numbers between 0 and 100.
If provided, the score for each corresponding pairs of quantiles will be calculated.
If not provided, it will default to every symmetrical pair of quantiles that can be found in \code{fcst},
ordered from widest to narrowest (e.x. the 25\% and 75\% quantiles are symmetrical).

The corresponding quantile intervals, if present, will be displayed in the resulting plot.

\code{quant_pairs} can be set to \code{list()} in order to display no quantile intervals.}

\item{score}{(Optional) A scoring function.
The function will be used to score \code{obs} against the forecast.
A scoring function should accept a forecast object, an observations data frame, as well as a \code{summarize} argument.
See \code{?accuracy}, \code{?neglog} for examples.
See \code{vignette(topic='casteval', package='casteval')} for details.}
}
\value{
A ggplot2 object
}
\description{
Plot a forecast, along with corresponding observations, quantile intervals, etc.
}
\examples{
fc <- create_forecast(list(
  time=1:3,
  vals=list(c(4,7,8), c(5,6,7), c(4,6,6))
))
obs <- data.frame(time=1:3, val_obs=5:7)

# plot forecast
plot_forecast(fc)

# plot forecast and observations
plot_forecast(fc, obs)

# plot forecast and confidence interval(s)
plot_forecast(fc, confs=c(50,95))

# highlight the observations inside the 95\% confidence interval
plot_forecast(fc, obs, confs=95, score=\(...) accuracy(..., interval=c(2.5, 97.5)))

# show the negative log score of each observation
plot_forecast(fc, obs, score=neglog)
}
