% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score.R
\name{plot_KDE}
\alias{plot_KDE}
\title{Plot the KDE used in \code{log_score}}
\usage{
plot_KDE(
  fcst,
  obs = NULL,
  at = NULL,
  after = NULL,
  bw = NULL,
  from = NULL,
  to = NULL,
  n = 101
)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame. Data from if will be included in the graph if provided}

\item{at}{(Optional) See \code{?log_score}}

\item{after}{(Optional) See \code{?log_score}}

\item{bw}{(Optional) See \code{?log_score}}

\item{from, to}{(Optional) The range over which the density will be plotted}

\item{n}{(Optional) How many points to calculate the density for}
}
\value{
A ggplot object
}
\description{
Create a plot displaying the density function calculated by the Kernel Density Estimation (KDE),
for the specified time in the forecast.
}
\examples{
#TODO
}
