% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean.R
\name{plot_mean}
\alias{plot_mean}
\title{Plot forecast mean}
\usage{
plot_mean(plt = NULL, fcst, alpha = 1, colour = "green")
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{alpha}{(Optional) The alpha parameter to be passed to \code{ggplot2}}

\item{colour}{(Optional) The colour to be passed to \code{ggplot2}}
}
\value{
A ggplot object
}
\description{
Given forecast data, plot the mean
}
\examples{
fc <- create_forecast(
  data.frame(time=1:5, val_mean=6:10)
)

NULL |> plot_mean(fc, colour="red", alpha=0.5)

fc2 <- create_forecast(
  data.frame(time=rep(1:5, each=3), val=c(1,2,3, 4,7,5, 10,11,10, 0,0,0, 1,2,2))
)

NULL |> plot_mean(fc2)
}
