% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observations.R
\name{plot_observations}
\alias{plot_observations}
\title{Intelligently plot observations}
\usage{
plot_observations(plt = NULL, obs, alpha = 0.4, colour = "black")
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{obs}{An observations data frame. If it contains a \code{score} column,
the observation points will be colour-coded according to the scores.
Otherwise, the \code{alpha} and \code{colour} parameters will determine the opacity and colour of the points.}

\item{alpha}{(Optional) The alpha parameter to be passed to \code{ggplot2}}

\item{colour}{(Optional) The colour to be passed to \code{ggplot2}}
}
\value{
A ggplot object.
}
\description{
Plot observation points. If scores are provided alongside the observations,
the plot can be colour-coded to convey information about score, accuracy, etc.
}
\examples{
fc <- create_forecast(list(
  time=1:3,
  vals=list(c(4,7,10), c(5,8,11), c(6,9,12))
))
obs <- data.frame(time=1:3, val_obs=c(5,9,13))

# plot observations on their own
plot_observations(NULL, obs)

# plot observations alongside forecast data
plot_observations(plot_ensemble(NULL, fc), obs)

# plot observations alongside forecast data, and colour-code by score
plot_observations(
  plot_ensemble(NULL, fc),
  log_score(fc, obs, summarize=FALSE)
)
}
