% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quantile}
\alias{plot_quantile}
\title{Plot a forecast quantile}
\usage{
plot_quantile(plt = NULL, fcst, quant, alpha = 0.5, colour = "orange")
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{quant}{A number between 0 and 100, the quantile to be graphed}

\item{alpha}{(Optional) The alpha parameter to be passed to \code{ggplot2}}

\item{colour}{(Optional) The colour to be passed to \code{ggplot2}}
}
\value{
desc
}
\description{
Given forecast data, plot one of its quantiles as a line
}
\examples{
#TODO
}
