% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot forecast quantiles}
\usage{
plot_quantiles(plt = NULL, fcst, quants, alpha = 0.5, colour = NULL)
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{quants}{A vector of numbers between 0 and 100, the quantiles to be graphed}

\item{alpha}{(Optional) The alpha parameter to be passed to \code{ggplot2}}

\item{colour}{(Optional) The colour to be passed to \code{ggplot2}}
}
\value{
desc
}
\description{
Given forecast data, plot the requested quantiles as lines
}
\examples{
#TODO
}
