% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_name}
\alias{quant_name}
\title{Turn quantile number into column name}
\usage{
quant_name(num)
}
\arguments{
\item{num}{The quantile number.}
}
\value{
The quantile column name.
}
\description{
Given a number, returns the column name of a corresponding quantile column
}
\examples{
# "val_q50"
casteval:::quant_name(50)

# "val_q2.5"
casteval:::quant_name(2.5)
}
