% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{quants2confs}
\alias{quants2confs}
\title{Get confidence intervals from quantiles}
\usage{
quants2confs(quants)
}
\arguments{
\item{quants}{Numeric vector containing quantile percentages.}
}
\value{
Numeric vector containing confidence intervals.
}
\description{
Helper function for graph_confidence_intervals().
Errors if the given quantiles are not sufficient to infer confidence intervals.
}
\examples{
# c(50,80,90,95,100)
casteval:::quants2confs(c(0,2.5,5,10,25,75,90,95,97.5,100))

# c(50,80)
casteval:::quants2confs(c(25,10,90,75))
}
