% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{raw2quant}
\alias{raw2quant}
\title{Compute quantile data from raw data}
\usage{
raw2quant(raw, perc)
}
\arguments{
\item{raw}{A numeric vector, or a list of non-empty numeric vectors. The vectors should not contain NA values.
Functions that use \code{raw2quant()} should make sure not to provide it NA values,
for example by using \code{remove_raw_NAs()}.}

\item{perc}{A percentage, from 0 to 100.}
}
\value{
A numeric vector with the same length as \code{raw}.
}
\description{
Given a column of raw data, calculate a given quantile.
}
\examples{
# c(2, 2, 4.5, 9)
casteval:::raw2quant(list(1:3, 2, 3:6, 7:11), 50)
# 26
casteval:::raw2quant(list(0:100), 26)
}
