% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{remove_raw_NAs}
\alias{remove_raw_NAs}
\title{Remove NA values from raw data}
\usage{
remove_raw_NAs(df)
}
\arguments{
\item{df}{The forecast data frame.
It should contain a \code{raw} column of numeric vectors.}
}
\value{
The data frame with raw NA values removed from the vectors.
}
\description{
Removes NA values from raw data in a forecast data frame.
Raises an error if this leaves any rows empty.
}
\examples{
# dplyr::tibble(time=1:3, raw=list(1, 1:3, 4))
casteval:::remove_raw_NAs(dplyr::tibble(time=1:3, raw=list(c(NA, 1), c(1, 2, 3), c(NA, NA, 4))))

# error if we end up with any empty rows
try(casteval:::remove_raw_NAs(dplyr::tibble(time=4:5, raw=list(c(1,NA), c(NA,NA)))))
}
