% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score forecasts}
\usage{
score(fcsts, obs, fun, ...)
}
\arguments{
\item{fcsts}{A single forecast object, or a list of forecast objects}

\item{obs}{An observations data frame}

\item{fun}{A scoring function. \code{fcst} and \code{obs} will be passed to it}

\item{...}{Additional parameters which will be passed along to \code{fun}}
}
\value{
If \code{fcsts} is a forecast object, a single score will be returned.
If \code{fcsts} is a list of forecast objects, then a list of scores will be returned.
}
\description{
Score a forecast against some observations using a given scoring function.
Additional parameters can be passed along to the scoring function.
}
\examples{
#TODO
}
