% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{validate_fcst_obs_pair}
\alias{validate_fcst_obs_pair}
\title{Validate a forecast-observations pair}
\usage{
validate_fcst_obs_pair(fcst, obs)
}
\arguments{
\item{fcst}{The forecast object (a named list, such as the output of \code{create_forecast()})}

\item{obs}{The observations (a data frame)}
}
\value{
NULL if valid. Error otherwise
}
\description{
Given a forecast and observations, verify that:
\itemize{
\item forecast is valid
\item observations are valid
\item forecast time type matches observations time type
}
}
\examples{
#TODO
}
