% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_forecast}
\alias{validate_forecast}
\title{Validate forecast in named-list format}
\usage{
validate_forecast(fcst)
}
\arguments{
\item{fcst}{The object to be validated}
}
\value{
NULL if valid (error if invalid)
}
\description{
\code{validate_forecast()} validates a forecast by checking that:
\itemize{
\item it is a named list
\item it contains a valid data frame
\item its forecast time is compatible with its time type
If any of these conditions fail, it stops with an appropriate an error message
}
}
\examples{
# valid forecast
casteval:::validate_forecast(list(
  name="hello",
  forecast_time=5,
  data=data.frame(time=1:3, val=4:6)
))

# not a forecast
try(casteval:::validate_forecast(data.frame(time=1:3, val=4:6)))

# bad forecast_time type
try(casteval:::validate_forecast(list(
  forecast_time=lubridate::as.Date(1000),
  data=data.frame(time=1:3, val=4:6)
)))
}
