% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_group_names}
\alias{validate_group_names}
\title{Validate forecast group names}
\usage{
validate_group_names(names)
}
\arguments{
\item{names}{A character vector of the group names (without the leading "grp_")}
}
\value{
NULL if valid, error otherwise
}
\description{
Check that the forecast groups don't contain any invalid names (in particular, the empty string)
}
\examples{
# valid
casteval:::validate_group_names(character(0))
casteval:::validate_group_names(c("variable", "scenario", "___12345"))

# invalid
try(
  casteval:::validate_group_names(c("variable", ""))
)
}
