% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{validate_interval}
\alias{validate_interval}
\title{Validate quantile interval vector}
\usage{
validate_interval(interval)
}
\arguments{
\item{interval}{Same as the \code{interval} parameter passed to accuracy()}
}
\value{
NULL if valid. Error otherwise.
}
\description{
Helper function for accuracy(). Performs input validation on its \code{interval} parameter.
}
\examples{
# valid
casteval:::validate_interval(c(50, 70))

# invalid
try(casteval:::validate_interval(c(70, 50)))

# invalid
try(casteval:::validate_interval(c(-1, 50)))

# invalid
try(casteval:::validate_interval(c(50,60,70)))

# invalid
try(casteval:::validate_interval("50, 60"))
}
