% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_obs}
\alias{validate_obs}
\title{Validate observations data frame}
\usage{
validate_obs(obs)
}
\arguments{
\item{obs}{An observations object.}
}
\value{
NULL if valid, error otherwise
}
\description{
Check that a given object:
\itemize{
\item is a data frame
\item is not empty
\item contains a valid \code{time} column
\item contains a numeric \code{val_obs} column
}
}
\examples{
# valid
casteval:::validate_obs(
  data.frame(time=1:3, val_obs=4:6)
)

# valid (contains `score` column)
casteval:::validate_obs(
  data.frame(time=1:3, val_obs=4:6, score=c(1, 0, -1))
)

# invalid
try(casteval:::validate_obs(
  data.frame(time=1:3, val_obs=c("a", "b", "c"))
))
}
