% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{validate_quant_interval}
\alias{validate_quant_interval}
\title{Validate quantile interval vector}
\usage{
validate_quant_interval(quants)
}
\arguments{
\item{quants}{Same as the \code{quants} parameter passed to accuracy()}
}
\value{
NULL if valid. Error otherwise.
}
\description{
Helper function for accuracy(). Performs input validation on its \code{quants} parameter.
}
\examples{
# valid
casteval:::validate_quant_interval(c(50, 70))

# invalid
try(casteval:::validate_quant_interval(c(70, 50)))

# invalid
try(casteval:::validate_quant_interval(c(-1, 50)))

# invalid
try(casteval:::validate_quant_interval(c(50,60,70)))

# invalid
try(casteval:::validate_quant_interval("50, 60"))
}
