% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quant.R
\name{validate_quant_name}
\alias{validate_quant_name}
\title{Validate quantile column name}
\usage{
validate_quant_name(name)
}
\arguments{
\item{name}{A string containing the quantile column name}
}
\value{
NULL if valid. Error otherwise
}
\description{
Check that a quantile column name starts with \code{val_q} followed by a number between 0 and 100.
}
\examples{
# valid
casteval:::validate_quant_name("val_q97.5")
casteval:::validate_quant_name("val_q50")
casteval:::validate_quant_name("val_q100")
casteval:::validate_quant_name("val_q0")

# invalid
try(casteval:::validate_quant_name("val_q50abc"))
try(casteval:::validate_quant_name("val_q101"))
try(casteval:::validate_quant_name("val_q-1"))
try(casteval:::validate_quant_name("hello"))
}
