% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_quant_order}
\alias{validate_quant_order}
\title{Make sure quantile values are logically possible}
\usage{
validate_quant_order(df)
}
\arguments{
\item{df}{A forecast data frame, with a \code{time} column and 0 or more \verb{quant_*} columns}
}
\value{
NULL if valid, error otherwise.
}
\description{
Checks that values in quantile columns are in increasing order for each time point.
That is, if \code{x < y}, then \code{df$quant_x[[i]] <= df$quant_y[[i]]} must hold true for all \verb{1 <= i <= nrow(df)}
}
\examples{
try(
  casteval:::validate_quant_order(data.frame(time=1:3, quant_25=4:6, quant_75=c(4,4,4)))
)
casteval:::validate_quant_order(data.frame(time=1:3, quant_2.5=4:6, quant_50=c(4,6,8)))
}
