% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quant.R
\name{validate_quant_pair}
\alias{validate_quant_pair}
\title{Validate quantile interval vector}
\usage{
validate_quant_pair(pair)
}
\arguments{
\item{pair}{An element of \code{quant_pairs}}
}
\value{
NULL if valid. Error otherwise.
}
\description{
Given a quantile pair, check that it is valid.
A valid quantile pair is a numeric vector of length 2
where both values are between 0 and 100,
and the first number is smaller than the second.
}
\examples{
# valid
casteval:::validate_quant_pair(c(50, 70))

# invalid
try(casteval:::validate_quant_pair(c(70, 50)))

# invalid
try(casteval:::validate_quant_pair(c(-1, 50)))

# invalid
try(casteval:::validate_quant_pair(c(50,60,70)))

# invalid
try(casteval:::validate_quant_pair("50, 60"))
}
