% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_forecast.R
\name{validate_time}
\alias{validate_time}
\title{Check that time compatible with forecast}
\usage{
validate_time(t, fcst)
}
\arguments{
\item{t}{A time (e.x. a number, date, or date-time).}

\item{fcst}{A forecast (as returned by \code{create_forecast()}).
No input validation is done on \code{fcst} itself.}
}
\value{
NULL if \code{t} is compatible with \code{fcst}. Error otherwise
}
\description{
Check that the type of a given time matches the time type of a given forecast.
}
\examples{
# both numeric (compatible)
casteval:::validate_time(5, create_forecast(data.frame(time=6,raw=7)))

# one date, one date-time (incompatible)
try(casteval:::validate_time(
  lubridate::ymd("2024-01-01"),
  create_forecast(data.frame(time=lubridate::ymd_hms("2024-01-01_00:00:00"),raw=6))
))
}
