% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_time_column}
\alias{validate_time_column}
\title{Validate time column}
\usage{
validate_time_column(times)
}
\arguments{
\item{times}{A vector, presumably a time column}
}
\value{
NULL if valid, error otherwise
}
\description{
Check that the type of a time column is valid.
}
\examples{
# valid
casteval:::validate_time_column(c(1,2,3))
casteval:::validate_time_column(lubridate::as_date(1000:1010))
casteval:::validate_time_column(lubridate::as_datetime(10000:10100))

# invalid
try(casteval:::validate_time_column(list(1,2,3)))
try(casteval:::validate_time_column(c("monday", "tuesday")))
}
