% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_ensemble.R
\name{wide2long}
\alias{wide2long}
\title{Convert raw forecast data to long format}
\usage{
wide2long(df)
}
\arguments{
\item{df}{The data frame. It should contain raw forecast data (see \code{create_forecast()}).}
}
\value{
The data frame in long format.
}
\description{
Given a data frame with raw data, convert it to long format.
Raw forecast data is usually stored in wide format, where the \code{raw} column
contains vectors of realization numbers.
In long format, each row contains a single data point for a single realization,
and a new column keeps track of the realization number for that row.
}
\examples{
# dplyr::tibble(
#   time=c(1,1,1,2,2,2,3,3,3),
#   realization=c(1,2,3,1,2,3,1,2,3),
#   raw=c(4,5,6,7,8,9,10,11,12)
# )
casteval:::wide2long(dplyr::tibble(
  time=1:3,
  raw=list(4:6, 7:9, 10:12)
))
}
