% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_frames.R
\name{widen_NULL}
\alias{widen_NULL}
\title{Replace NULLs with vectors of NAs}
\usage{
widen_NULL(lst, len)
}
\arguments{
\item{lst}{A list of vectors and/or NULLs.
All NULLs in the list will be replaced with a vector of \code{len} NAs}

\item{len}{A positive integer.}
}
\value{
The modified \code{lst}.
}
\description{
Helper for \code{combine_two_data_frames()}.
Replaces NULLs in a list with vectors of NAs of the correct length
}
\examples{
casteval:::widen_NULL(list(1, 2, NULL, 4, NULL), 1)
casteval:::widen_NULL(list(c(1,2,3), c(4,5,6), NULL), 3)
}
