% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Get accuracy from quantiles}
\usage{
accuracy(fcst, obs, interval = NULL)
}
\arguments{
\item{fcst}{The forecast (see \code{create_forecast()} output).
If quantiles are provided, they will be used to compute the accuracy.
If raw data is provided, quantiles will be calculated according to the \code{interval} parameter.}

\item{obs}{The observations data frame.}

\item{interval}{Either NULL or a vector of two numbers from 0 to 100.
If \code{fcst} contains quantile data then the corresponding quantile columns will be
used as the confidence interval.
If \code{fcst} contains raw data then the corresponding quantiles will be calculated and
used as a confidence interval.}
}
\value{
A number from 0 to 1,
the rate at which the observations were inside the specified confidence interval
}
\description{
Given a forecast and set of observations,
compute the accuracy (# within confidence interval / # total) of the forecast.
}
\examples{
#TODO
}
