% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_facets}
\alias{apply_facets}
\title{Apply facets to plot}
\usage{
apply_facets(plot, plotting_groups)
}
\arguments{
\item{plot}{A \code{ggplot} object}

\item{plotting_groups}{A character vector of up to 2 group column names}
}
\value{
desc
}
\description{
Apply the given facets to the given plot
}
\details{
\itemize{
\item If \code{plotting_groups} is empty, \code{plot} is returned unchanged
\item If \code{plotting_groups} has one element, \code{ggplot2::facet_wrap()} is used
\item If \code{plotting_groups} has two elements, \code{ggplot2::facet_grid()} is used
}
}
\examples{
obs <- groups_obs |> dplyr::filter(grp_scenario==1)
NULL |> plot_observations(obs) |> casteval:::apply_facets(c("grp_variable"))
NULL |> plot_observations(obs) |> casteval:::apply_facets(c("grp_variable", "grp_province"))
}
