% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Calculate forecast bias}
\usage{
bias(fcst, obs, summarize = TRUE)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{An observations data frame.}

\item{summarize}{A boolean, defaults to TRUE. If TRUE, a single number will be returned as the score for the forecast.
If FALSE, a data frame with \code{time} and \code{score} columns (and possibly others) will be returned,
containing the intermediate values for each time point used to calculate the score.
The type of the \code{score} column can be anything (numbers, booleans, etc.).
This can be used by graphing functions to color code observations, for example.}
}
\value{
A number between -1 and 1, inclusive. -1 means 100\% underprediction and 1 means 100\% overprediction.
}
\description{
Calculate forecast bias
}
\examples{
#TODO
}
