% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casteval-package.R
\docType{package}
\name{casteval-package}
\alias{casteval}
\alias{casteval-package}
\title{casteval: Modular tools for scoring and comparing infectious disease forecasts}
\description{
Provides a generalized, modular set of tools to facilitate/automate the scoring and comparing of infectious disease forecasts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://phrsd.pages.cscscience.ca/risk/casteval}
}

}
\author{
\strong{Maintainer}: Daniel Yu \email{daniel.yu@phac-aspc.gc.ca}

Authors:
\itemize{
  \item Irena Papst \email{irena.papst@phac-aspc.gc.ca}
}

}
\keyword{internal}
