% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_frames.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combine a group of forecast data frames into one}
\usage{
combine_data_frames(dfs)
}
\arguments{
\item{dfs}{A list of forecast data frames, with raw data points and the same time types.}
}
\value{
A data frame containing the aggregated data.
}
\description{
\code{combine_data_frames()} converts a list of forecast data frames
into a single data frame containing all the aggregated data points in one column.
It performs no input validation so it is not exported.
}
\examples{
# missing values are filled in with NA
casteval:::combine_data_frames(list(
  data.frame(time=c(1,2,3), raw=c(10,11,12)),
  data.frame(time=c(2,3,4), raw=c(20,21,22)),
  data.frame(time=c(3,4,5), raw=c(30,31,32))
))
}
