% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast}
\alias{create_forecast}
\title{Create a forecast}
\usage{
create_forecast(data, name = NULL, forecast_time = NULL)
}
\arguments{
\item{data}{Forecast data. Either a data frame or list of data frames.
An ensemble of forecasts can be provided as a list of data frames, in which case it will be aggregated into a single data frame.
The data frame(s) should contain a column named \code{time}, which may be of strings, integers, dates, or date-times.
No mixing of types is allowed (e.g. \code{time} may not contain both integers and date-times).}

\item{name}{A string specifying the name of the forecast/model.}

\item{forecast_time}{An integer, date, or date-time specifying when the forecast was created.
Its type should match the type of values in the \code{time} column(s) of \code{data}
If provided, this forecast will be scored only using data corresponding to dates/times greater than or equal to \code{forecast_time}.
Additionally, graphs of this forecast will highlight the \code{forecast_time} using a vertical line.}
}
\value{
A named list containing the forecast and its metadata
}
\description{
\code{create_forecast()} creates a forecast object given data and optional metadata.
It accepts a variety of forecast formats as input and intelligently converts them into a standardized format.
}
\examples{
# TBD
}
