% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast}
\alias{create_forecast}
\title{Create a forecast}
\usage{
create_forecast(dat, name = NULL, forecast_time = NULL)
}
\arguments{
\item{dat}{Forecast data. It can be one of the following:
\itemize{
\item A single data frame containing raw or summary data.
\item A list of data frames each containing raw data.
\item A named list containing a numeric vector \code{time} and a list of vectors \code{data}.
Each vector corresponds to a realization in a forecast ensemble.
The \code{time} vector and \code{data} vectors must all have the same length.
}

All data frames in \code{dat} must additionally have a \code{time} column containing either integers, dates, or date-times.

Raw data columns should be named \code{raw}.
Summary data columns may be named \code{mean}, or \code{quant_} followed by a number from 0 to 100 (e.x. \code{quant_2.5}, \code{quant_50}, etc.).
All raw or summary data must be numeric.

See below for examples.

See \code{\link[=get_format]{get_format()}} for further details on data frame formatting.}

\item{name}{(Optional) A string specifying the name of the forecast/model.}

\item{forecast_time}{(Optional) An integer, date, or date-time specifying when the forecast was created.
Its type should match the type of values in the \code{time} column(s) of \code{data}
If provided, this forecast will be scored only using data corresponding to dates/times greater than or equal to \code{forecast_time}.
Additionally, plots of this forecast may graphically distinguish between values to the left and right of \code{forecast_time}.}
}
\value{
A named list containing the forecast and its metadata.
The (possibly processed) data frame is stored in \verb{$data}.
The name and forecast time are stored in \verb{$name} and \verb{$forecast_time}.
The type of the time column (one of "date", "date-time", or "numeric") is stored in \verb{$time_type}
The types of the data columns (a character vector containing "mean", "quant", and/or "raw") are stored in \verb{$data_types}
}
\description{
\code{create_forecast()} creates a forecast object given data and optional metadata.
It accepts a variety of forecast formats as input and intelligently converts them into a standardized format.
}
\examples{
# forecast with numeric times and raw data
create_forecast(data.frame(time=1:3, raw=10:12), name="a forecast", forecast_time=2)

# forecast with dates and mean-and-quantiles data
create_forecast(
  data.frame(
    time=c(lubridate::ymd("2024-01-01"), lubridate::ymd("2024-01-02")),
    mean=10:11, quant_2.5=5:6, quant_97.5=15:16
  ),
  name="another forecast"
)

# combining an ensemble of realizations into one
create_forecast(list(
  dplyr::tibble(time=1:5,raw=6:10),
  dplyr::tibble(time=2:6,raw=7:11),
  dplyr::tibble(time=3:7,raw=8:12)
))

# an already-combined ensemble
create_forecast(dplyr::tibble(time=1:2, raw=list(10:15, 20:25)))
}
