% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast.R
\name{create_forecast_ensemble}
\alias{create_forecast_ensemble}
\title{Create forecast from time vector and ensemble of realizations}
\usage{
create_forecast_ensemble(dat, name, forecast_time)
}
\arguments{
\item{dat}{A named list containing names \code{time} and \code{ensemble}}

\item{name}{A string}

\item{forecast_time}{A number, date, or date-time}
}
\value{
A forecast object
}
\description{
Helper for \code{create_forecast()}.
}
\examples{
# See `create_forecast()`
}
