% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\name{crps}
\alias{crps}
\title{Compute Continuous Ranked Probability Score for forecast}
\usage{
crps(fcst, obs, at = NULL, after = NULL, summarize = TRUE)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{An observations data frame.}

\item{at}{(Optional) A time (must be compatible with \code{fcst} and \code{obs}).
If specified, the score for this time point will be returned.
Mutually exclusive with \code{after}.}

\item{after}{(Optional) A number. If specified, the score at
time \code{fcst$forecast_time + after} will be returned.
Mutually exclusive with \code{at}.}

\item{summarize}{A boolean, defaults to TRUE. If TRUE, a single number will be returned as the score for the forecast.
If FALSE, a data frame with columns named \code{time}, \code{val_obs}, and \code{score} will be returned,
containing the scores for each individual time point.
This can be used by plotting functions to colour-code observations, for example.}
}
\value{
If \code{summarize} is \code{FALSE},
a data frame containing times, observations, and scores.
Otherwise, the score at the time speficied by either \code{at} or \code{after}.
}
\description{
Given a forecast and set of observations, compute the
}
\details{
Continuous Ranked Probability Score (CRPS) for every time point.
The CRPS for a given distribution \code{f} and observation \code{y},
let \code{F} be the CDF of \code{f}. Then the CRPS is the integral of the square
of \code{F(x) - H(x - y)}, where \code{H} is the Heaviside function.

The CDF is obtained using an emperical distribution function on the forecast data.
}
\examples{
fc <- create_forecast(list(
  time=c(1,2,3),
  vals=list(c(1,1,1), c(2,2,2), c(3,3,3), c(4,4,4), c(5,5,5)), forecast_time=1)
)

crps(fc, data.frame(time=1:3, val_obs=c(3,4,5)), summarize=FALSE)

crps(fc, data.frame(time=1:3, val_obs=c(3,4,5)), at=2)
}
