% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020fc}
\alias{denmark2020fc}
\title{Denmark 2020 COVID-19 forecast object}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://www.nature.com/articles/s41567-020-01121-y}

\url{https://github.com/jonassjuul/curvestat}
}
\usage{
denmark2020fc
}
\description{
A \code{{casteval}} forecast object containing data from a Danish COVID-19 forecasting effort,
concerning 5 May to 1 October 2020.
}
\seealso{
\link{denmark2020df}, \link{denmark2020obs}, \link{denmark2020ens}
}
\keyword{datasets}
