% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020fc}
\alias{denmark2020fc}
\title{Denmark 2020 COVID-19 forecast data}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://www.nature.com/articles/s41567-020-01121-y}

\url{https://github.com/jonassjuul/curvestat}
}
\usage{
denmark2020fc
}
\description{
Forecast realizations from a Danish COVID-19 forecasting effort,
concerning 5 May to 1 October 2020.
}
\details{
Formatted as a \code{casteval} forecast object.

Acquired from the below paper/repository.
}
\keyword{datasets}
