% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark2020obs}
\alias{denmark2020obs}
\title{Denmark 2020 COVID-19 observations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 2 columns.
}
\source{
\url{https://www.nature.com/articles/s41567-020-01121-y}

\url{https://github.com/jonassjuul/curvestat}
}
\usage{
denmark2020obs
}
\description{
Daily COVID-19 hospitalizations in Denmark from the 55 days leading up to
the 5 May 2020 forecast (see \code{?denmark2020fc}), from the same source.
}
\details{
Formatted as a two-column data frame.
}
\keyword{datasets}
