% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_format}
\alias{get_format}
\title{Get data frame format}
\usage{
get_format(df)
}
\arguments{
\item{df}{A data frame. It should have a valid time column and a valid combination of data columns.}
}
\value{
A named list containing a \code{time_type} (a string containing the type of the values in the time column)
and \code{data_type} (a character vector containing the types of data columns present)
}
\description{
Helper function for \code{create_forecast()}
\code{get_format()} inspects the columns of a data frame to determine how it's formatted.
Also performs input validation on the contents of the data frame.
}
\examples{
# TBD
}
