% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{get_plotting_groups}
\alias{get_plotting_groups}
\title{Find the groups relevant to plotting}
\usage{
get_plotting_groups(df)
}
\arguments{
\item{df}{A forecast data frame}
}
\value{
A character vector containing all the group column names with multiple values
}
\description{
\code{{casteval}} allows plotting of forecasts with up to 2 groups.
This function finds all group columns with more than one value in them.
}
\examples{
casteval:::get_plotting_groups(data.frame(time=1, val=2))

casteval:::get_plotting_groups(data.frame(time=1:3, val=4:6, grp_var=7:9, grp_loc=8:10))

casteval:::get_plotting_groups(data.frame(time=1:3, val=4:6, grp_var=7:9, grp_loc=10:12, grp_sce=13:15))

casteval:::get_plotting_groups(data.frame(time=1:3, val=4:6, grp_var=7:9, grp_loc=c(1,1,1), grp_sce=13:15))
}
