% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{get_quantile}
\alias{get_quantile}
\title{Obtain quantiles from data frame}
\usage{
get_quantile(df, perc)
}
\arguments{
\item{df}{A forecast data frame. It should contain raw or quantile data.}

\item{perc}{A percentage, from 0 to 100.}
}
\value{
A data frame with a \code{time} column and a \code{quant} column, containing the requested quantile data
}
\description{
Given a forecast data frame, compute quantiles from raw data or existing quantile columns,
depending on what is present.
}
\examples{
df1 <- data.frame(time=rep(1:3,each=3), sim=rep(1:3, 3), val=c(4:6, 7:9, 10:12))
df2 <- data.frame(time=1:3, val_q50=c(1000, 2000, 3000))

casteval:::get_quantile(df1, 50)
casteval:::get_quantile(df2, 50)
}
