% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_type}
\alias{get_time_type}
\title{Get type of data frame time column}
\usage{
get_time_type(df)
}
\arguments{
\item{df}{A data frame with a \code{time} column.}
}
\value{
A string, one of "numeric", "date", "date-time"
}
\description{
Inspects a data frame's time column and returns its type.
}
\examples{
# "numeric"
casteval:::get_time_type(data.frame(time=c(1,2,3)))

# "date"
casteval:::get_time_type(data.frame(time=lubridate::as_date(1:3)))

# "date-time"
casteval:::get_time_type(data.frame(time=lubridate::ymd_hms("2024-01-01_12:34:56")))

# unsupported type
try(casteval:::get_time_type(list("January 1", "January 2")))
}
