% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_confidence_intervals}
\alias{graph_confidence_intervals}
\title{Graph forecast confidence intervals}
\usage{
graph_confidence_intervals(graph = NULL, fcst, confs = NULL)
}
\arguments{
\item{graph}{A ggplot object, or \code{NULL}. If not \code{NULL},
it will be added to using \code{ggplot2}'s \code{+} operator.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{confs}{A numeric vector containing the confidence intervals to be displayed, as percentages.
If \code{NULL}, the confidence intervals will be inferred from the quantiles present in the forecast data frame.}
}
\value{
A ggplot object.
}
\description{
Graph shaded bands indicating confidence intervals of a forecast.
}
\examples{
fc <- create_forecast(dplyr::tibble(
  time=1:3,
  raw=list(c(3,5,6,7,3), c(6,8,7,8,7), c(11,15,13,14,17))
))
NULL |> casteval:::graph_ensemble(fc) |> casteval:::graph_confidence_intervals(fc, c(90,50))
}
