% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_ensemble.R
\name{graph_ensemble}
\alias{graph_ensemble}
\title{Graph ensemble of forecast realizations}
\usage{
graph_ensemble(graph = NULL, fcst, alpha = NULL)
}
\arguments{
\item{graph}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{alpha}{(Optional) The alpha value to be passed to \code{ggplot2::geom_line()}.}
}
\value{
A ggplot object.
}
\description{
Given a forecast with raw realizations, generate a graph displaying all of them.
}
\examples{
casteval:::graph_ensemble(
  NULL,
  create_forecast(dplyr::tibble(time=1:3, raw=list(4:6, 7:9, 10:12))
))

casteval:::graph_ensemble(
  NULL,
  create_forecast(data.frame(time=lubridate::as_datetime(c(0,20000,100000)), raw=c(20,30,40))
))
}
