% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_forecast}
\alias{graph_forecast}
\title{Graph a forecast}
\usage{
graph_forecast(fcst, obs = NULL, raw = TRUE, confs = NULL, score = NULL)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame.
If provided, they will be overlaid over the forecast as points.}

\item{raw}{(Optional) A boolean. Defaults to TRUE, in which case raw ensemble curves will be displayed.}

\item{confs}{(Optional) A vector of numbers from 0 to 100.
The corresponding confidence interval(s) will be displayed in the resulting graph.}

\item{score}{(Optional) A scoring function.
The function will be used to score the \code{obs} against the forecast.
A scoring function should accept a forecast object, an observations data frame, as well as a \code{summarize} argument.
See \code{accuracy()}, \code{neglog()} for examples}
}
\value{
desc
}
\description{
Graph a forecast, along with corresponding observations, confidence intervals, etc.
}
\examples{
#TODO
}
