% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_forecast}
\alias{graph_forecast}
\title{Graph a forecast}
\usage{
graph_forecast(fcst, obs = NULL, confs = NULL, score = NULL)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame.
If provided, they will be overlaid over the forecast as points.}

\item{confs}{(Optional) A vector of numbers from 0 to 100.
The corresponding confidence interval(s) will be displayed in the resulting graph.}

\item{score}{(Optional) A scoring function.
The function will be used to score the \code{obs} against the forecast.
A scoring function should accept a forecast object, an observations data frame, as well as a \code{summarize} argument.
See \code{accuracy()}, \code{neglog()} for examples}
}
\value{
desc
}
\description{
Graph a forecast, along with corresponding observations, confidence intervals, etc.
}
\examples{
fc <- create_forecast(dplyr::tibble(
  time=1:3,
  raw=list(c(4,5,4), c(7,6,6), c(8,7,6))
))
obs <- data.frame(time=1:3, obs=5:7)

# graph forecast
graph_forecast(fc)

# graph forecast and observations
graph_forecast(fc, obs)

# graph forecast and confidence interval(s)
graph_forecast(fc, confs=c(50,95))

# highlight the observations inside the 95\% confidence interval
graph_forecast(fc, obs, confs=95, score=\(...) accuracy(..., interval=c(2.5, 97.5)))

# show the negative log score of each observation
graph_forecast(fc, obs, score=neglog)
}
