% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_observations.R
\name{graph_observations}
\alias{graph_observations}
\title{Intelligently graph observations}
\usage{
graph_observations(graph = NULL, obs)
}
\arguments{
\item{graph}{A ggplot object, or \code{NULL}. If not \code{NULL},
it will be added to using \code{ggplot2}'s \code{+} operator.}

\item{obs}{An observations data frame. If it contains a \code{score} column,
the observation points will be color-coded according to the scores.}
}
\value{
A ggplot object.
}
\description{
Graph observation points. If provided with a forecast,
the points can be color-coded to convey information about score, accuracy, etc.
}
\examples{
#TODO
}
