% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_observations.R
\name{graph_observations}
\alias{graph_observations}
\title{Intelligently graph observations}
\usage{
graph_observations(graph = NULL, obs, fcst = NULL)
}
\arguments{
\item{name}{desc}
}
\value{
desc
}
\description{
Graph observation points. If provided with a forecast,
the points can be color-coded to convey information about score, accuracy, etc.
}
\examples{
#TODO
}
