% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_quantiles.R
\name{graph_quantiles}
\alias{graph_quantiles}
\title{Graph forecast quantiles}
\usage{
graph_quantiles(graph = NULL, fcst, quants = NULL)
}
\arguments{
\item{graph}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{quants}{A numeric vector containing the quantiles to be graphed, as percentages.
If \code{NULL}, all quantiles present in the forecast data frame will be graphed.}
}
\value{
A ggplot object.
}
\description{
Graph lines indicating quantiles of a forecast.
}
\examples{
fc1 <- create_forecast(dplyr::tibble(
time=1:3,
raw=list(0:5, 5:10, 10:15)
))

NULL |> casteval:::graph_ensemble(fc1) |> casteval:::graph_quantiles(fc1, c(2.5, 25,51,75))
}
