% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_all}
\alias{group_all}
\title{Group forecast data frame}
\usage{
group_all(df, .add = FALSE)
}
\arguments{
\item{df}{A forecast data frame}

\item{.add}{(Optional) The \code{.add} parameter to be passed to \code{dplyr::group_by()}.
Defaults to \code{FALSE}}
}
\value{
The grouped forecast data frame
}
\description{
Given a forecast data frame, group it by all the \verb{grp_*} columns
}
\examples{
# `{casteval}` exports an example data frame for grouping
casteval:::group_all(groupex)

dplyr::group_by(groupex, time) |> casteval:::group_all(.add=TRUE)
}
