% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_all}
\alias{group_all}
\title{Group forecast data frame}
\usage{
group_all(df, ...)
}
\arguments{
\item{df}{A forecast data frame}

\item{...}{Additional arguments to be passed to \code{dplyr::group_by()} (e.x. \code{.add}, \code{.drop})}
}
\value{
The grouped forecast data frame
}
\description{
Given a forecast data frame, group it by all its \verb{grp_*} columns.
}
\examples{
# `{casteval}` exports an example data frame for grouping
casteval:::group_all(groups1)

dplyr::group_by(groups1, time) |> casteval:::group_all(.add=TRUE)
}
