% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupex_obs}
\alias{groupex_obs}
\title{Grouping example observations
An example observations data frame for demonstration \code{{casteval}}'s grouping functionality.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 5 columns.
}
\usage{
groupex_obs
}
\description{
Grouping example observations
An example observations data frame for demonstration \code{{casteval}}'s grouping functionality.
}
\keyword{datasets}
