% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups1}
\alias{groups1}
\title{Grouping example forecast 1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 7 columns.
}
\usage{
groups1
}
\description{
An example forecast data frame for demonstrating (and testing) \code{{casteval}}'s grouping functionality.
Contains mean-and-quantiles data.
}
\keyword{datasets}
