% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups2}
\alias{groups2}
\title{Grouping example forecast 2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 5 columns.
}
\usage{
groups2
}
\description{
An example forecast data frame for demonstrating (and testing) \code{{casteval}}'s grouping functionality.
Contains raw data.
}
\keyword{datasets}
