% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Integer breaks on ggplot2 axes}
\source{
\url{https://gist.github.com/jhrcook/eb7b63cc57c683a6eb4986c4107a88ec}
}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Number of breaks.}

\item{...}{other arguments passed to \code{pretty()}}
}
\description{
A modified version of \code{scales::pretty_breaks()} for only integer-valued ticks.
}
