% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_data}
\alias{join_data}
\title{Join two data frames}
\usage{
join_data(df1, df2)
}
\arguments{
\item{df1}{The first data frame}

\item{df2}{The second data frame}
}
\value{
The inner join of the two data frames
}
\description{
Join two data frames by time and grouping columns.
}
\details{
\code{df1} and \code{df2} should both have a \code{time} column and the exact same grouping columns.
}
\examples{
#TODO
}
