% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{join_fcst_obs}
\alias{join_fcst_obs}
\title{Join a forecast and observations into a single data frame}
\usage{
join_fcst_obs(df, obs)
}
\arguments{
\item{df}{The forecast data frame.}

\item{obs}{The observations data frame.}
}
\value{
The forecast data frame with an additional \code{obs} column containing observations.
}
\description{
Add an observations column to the forecast data frame containing
observations for each time point.
}
\examples{
# data.frame(time=1:3, raw=4:6, obs=8:10)
casteval:::join_fcst_obs(data.frame(time=1:3, val=4:6), data.frame(time=0:4, val_obs=7:11))

# remove rows with missing observations
# data.frame(time=3, val_q50=6, val_obs=7)
casteval:::join_fcst_obs(
  data.frame(time=1:3, val_q50=4:6),
  data.frame(time=2:3, val_obs=c(NA,7))
)

# default behaviour is to error if observations are missing
try(casteval:::join_fcst_obs(
  data.frame(time=1:3, val_q50=4:6),
  data.frame(time=2:3, val_obs=c(NA,7))
))
}
