% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_score.R
\name{make_log_score}
\alias{make_log_score}
\title{\code{log_score()} function factory}
\usage{
make_log_score(at = NULL, after = NULL, bw = NULL)
}
\arguments{
\item{at}{See \code{?log_score}}

\item{after}{See \code{?log_score}}

\item{bw}{See \code{?log_score}}
}
\value{
A function wrapping \code{log_score()} with the given parameters
}
\description{
Create a function wrapping \code{log_score()} with specified parameters.
}
\examples{
#TODO
}
