% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KDE.R
\name{plot_KDE}
\alias{plot_KDE}
\title{Plot the KDE used in \code{log_score()}}
\usage{
plot_KDE(
  fcst,
  obs = NULL,
  at = NULL,
  after = NULL,
  bw = NULL,
  from = NULL,
  to = NULL,
  n = 101,
  binwidth = NULL
)
}
\arguments{
\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{(Optional) An observations data frame. Data from if will be included in the graph if provided}

\item{at}{(Optional) See \code{?log_score}}

\item{after}{(Optional) See \code{?log_score}}

\item{bw}{(Optional) See \code{?log_score}}

\item{from, to}{(Optional) The range over which the density will be plotted}

\item{n}{(Optional) How many points to calculate the density for}

\item{binwidth}{(Optional) The binwidth for the histogram. Not to be confused with \code{bw}, which stands for bandwidth and is unrelated.}
}
\value{
A ggplot object
}
\description{
Create a plot displaying the density function calculated by the Kernel Density Estimation (KDE),
for the specified time in the forecast.
}
\details{
Creates a plot containing:
\itemize{
\item The density curve at the given time calculated by the KDE
\item The forecast data points at the given time, along the x-axis
\item A histogram of the density of the forecast data points
\item A vertical line showing the observation at the given time (if provided)
}
}
\examples{
withr::with_seed(42, {
  dat <- rnorm(100)
  fc <- create_forecast(dplyr::tibble(time=rep(1,100), val=dat), forecast_time=1)
  obs <- data.frame(time=1, val_obs=1)
  plot_KDE(fc, obs, at=1)
  plot_KDE(fc, obs, at=1, bw=0.1)
  plot_KDE(fc, obs, at=1, bw=1)
})
}
