% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observations.R
\name{plot_obs_score}
\alias{plot_obs_score}
\title{Plot and score observations}
\usage{
plot_obs_score(plt = NULL, fcst, obs, invert_scale = FALSE, score, ...)
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{obs}{An observations data frame}

\item{invert_scale}{(Optional) a boolean. If \code{TRUE}, the color scale will be inverted.
This is useful for scores where smaller values are better, e.x. CRPS.}

\item{score}{(Optional) A scoring function.
The function will be used to score \code{fcst} against \code{obs}.
A scoring function should accept a forecast object, an observations data frame, as well as a \code{summarize} argument.
See \code{?accuracy}, \code{?log_score} for examples.
See \code{vignette(topic='casteval', package='casteval')} for details.}

\item{...}{Additional parameters to be passed to \code{score}.
Note that \code{summarize} should not be one of them,
since \code{casteval} already passes that to \code{score}.}
}
\value{
A ggplot object
}
\description{
Wrapper for \code{plot_observations()}.
Plots scores the given forecast against the given observations,
then plots the observations with a colour scale corresponding to score.
}
\examples{
#TODO
}
