% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quant_intervals}
\alias{plot_quant_intervals}
\title{Plot quantile intervals}
\usage{
plot_quant_intervals(
  plt = NULL,
  fcst,
  quant_pairs = NULL,
  alpha = NULL,
  palette = 1
)
}
\arguments{
\item{plt}{(Optional) A ggplot object.
it will be added to using the \code{+} operator.
Defaults to NULL, in which case a blank ggplot object will be created.}

\item{fcst}{A forecast object (see output of \code{create_forecast()}).}

\item{quant_pairs}{(Optional) A list of pairs of numbers between 0 and 100.
If not provided, the quantile intervals will be inferred from the forecast.
Otherwise, each pair will be used as a quantile intervals.}

\item{alpha}{(Optional) The alpha parameter to be passed to \code{ggplot2}}

\item{palette}{(Optional) The colour palette to use (see \code{?ggplot2::scale_fill_brewer})}
}
\value{
A \code{ggplot2} object.
}
\description{
Given forecast data, highlight the specified quantile intervals.
}
\examples{
#TODO
}
