% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score forecasts}
\usage{
score(fcsts, obs, fun, ...)
}
\arguments{
\item{fcsts}{A single forecast object, or a list of forecast objects}

\item{obs}{An observations data frame}

\item{fun}{A scoring function. \code{fcst} and \code{obs} will be passed to it}

\item{...}{Additional parameters which will be passed along to \code{fun}}
}
\value{
If \code{fcsts} is a forecast object, a single score will be returned.
If \code{fcsts} is a list of forecast objects, then a list of scores will be returned.
}
\description{
Score a forecast against some observations using a given scoring function.
Additional parameters can be passed along to the scoring function.
}
\details{
See \code{vignette("scoring")} for a complete list of scoring functions available, along with the details of their use.
}
\examples{
fc1 <- create_forecast(data.frame(time=1:5, val=6:10))
fc2 <- create_forecast(data.frame(time=1:5, val=7:11))

obs <- data.frame(time=1:5, val_obs=c(8,8,8,8,8))

score(fc1, obs, bias)
score(fc2, obs, bias)
score(list(fc1, fc2), obs, bias)
}
