% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_column}
\alias{validate_column}
\title{Check that column is in data frame}
\usage{
validate_column(df, col)
}
\arguments{
\item{df}{A data frame.}

\item{col}{A string containing the column name.}
}
\value{
NULL if valid, error otherwise.
}
\description{
Errors of \code{col} is not a column in data frame.
}
\examples{
casteval:::validate_column(
  data.frame(time=1, raw=4, quant_25=6, quant_75=8),
  "quant_25"
)
try(casteval:::validate_column(
  data.frame(time=1, raw=4),
  "mean"
))
}
