% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate forecast data frame}
\usage{
validate_data_frame(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
NULL on success. Error otherwise
}
\description{
Given a data frame, checks that:
}
\details{
\itemize{
\item it isn't empty
\item it has a valid time column
\item its quantile columns are properly named and their values are in order
\item it has at least one data column
\item its data columns are numeric
}
}
\examples{
# TODO
}
