% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{validate_fcst_obs_pair}
\alias{validate_fcst_obs_pair}
\title{Validate a forecast-observations pair}
\usage{
validate_fcst_obs_pair(fcst, obs)
}
\arguments{
\item{fcst}{The forecast object (a named list, such as the output of \code{create_forecast()})}

\item{obs}{The observations (a data frame)}
}
\value{
NULL if valid. Error otherwise
}
\description{
Given a forecast and observations, verify that:
\itemize{
\item forecast is valid
\item observations are valid
\item forecast time type matches observations time type
}
}
\examples{
# compatible time types
casteval:::validate_fcst_obs_pair(
  create_forecast(data.frame(time=1:10, raw=11:20)),
  data.frame(time=101:110, obs=111:120)
)

# incompatible time types
try(casteval:::validate_fcst_obs_pair(
  create_forecast(data.frame(time=1:10, raw=11:20)),
  data.frame(time=lubridate::ymd("2024-01-01"), obs=5)
))
}
