% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_forecast}
\alias{validate_forecast}
\title{Validate forecast in named-list format}
\usage{
validate_forecast(fcst)
}
\arguments{
\item{fcst}{The object to be validated}
}
\value{
NULL if valid (error if invalid)
}
\description{
\code{validate_forecast()} validates a forecast by checking that it:
\itemize{
\item is a named list
\item contains a \code{time_type} field
\item contains a \code{data_types} field
\item contains a valid data frame
\item data frame format matches \code{time_type} and \code{data_type}
If any of these conditions fail, it stops with an appropriate an error message
}
}
\examples{
# incorrect time type
try(casteval:::validate_forecast(list(
  time_type="date", data_types="raw", data=data.frame(time=1:3, raw=4:6)
)))

# invalid `forecast_time` type
try(casteval:::validate_forecast(list(
  time_type="numeric",
  data_types="raw",
  data=data.frame(time=1,raw=4),
  forecast_time=lubridate::ymd("2024-01-01")
)))

# a valid forecast
casteval:::validate_forecast(list(
  time_type="numeric",
  data_types=c("mean", "quant"),
  data=data.frame(time=1:3, mean=4:6, quant_50=7:9),
  forecast_time=2
))
}
