% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_quant_name}
\alias{validate_quant_name}
\title{Validate quantile column name}
\usage{
validate_quant_name(name)
}
\arguments{
\item{name}{A string containing the quantile column name}
}
\value{
NULL if valid. Error otherwise
}
\description{
Check that a quantile column name starts with \code{val_q} followed by a number between 0 and 100.
}
\examples{
#TODO
}
