% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_time_column}
\alias{validate_time_column}
\title{Validate time column}
\usage{
validate_time_column(times)
}
\arguments{
\item{times}{A vector, presumably a time column}
}
\value{
NULL if valid, error otherwise
}
\description{
Check that the type of a time column is valid.
}
\examples{
#TODO
}
