% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ensemble.R
\name{widen_NULL}
\alias{widen_NULL}
\title{Replace NULLs with vectors of NAs}
\usage{
widen_NULL(lst, len)
}
\arguments{
\item{lst}{A list of vectors and/or NULLs.
All NULLs in the list will be replaced with a vector of \code{len} NAs}

\item{len}{A positive integer.}
}
\value{
The modified \code{lst}.
}
\description{
Helper for \code{combine_two_data_frames()}.
Replaces NULLs in a list with vectors of NAs of the correct length
}
\examples{
# TBD
}
